chore(dashboard): use dynamic imports and log build
