+++ frontend/admin-dashboard/src/pages/dashboard/roles.tsx
@@ -1,36 +1,16 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import type { GetStaticProps } from 'next';
 import { useTranslation } from 'react-i18next';
+import dynamic from 'next/dynamic';
 
-interface Assignment {
-  username: string;
-  role: string;
-}
+const RolesList = dynamic(() => import('../../components/RolesList'));
 
 export default function RolesPage() {
   const { t } = useTranslation();
-  const [roles, setRoles] = useState<Assignment[]>([]);
-  useEffect(() => {
-    async function load() {
-      const resp = await fetch('/roles', {
-        headers: { Authorization: 'Bearer ' + localStorage.getItem('token') },
-      });
-      if (resp.ok) {
-        setRoles(await resp.json());
-      }
-    }
-    void load();
-  }, []);
   return (
     <div>
       <h1>{t('roles')}</h1>
-      <ul>
-        {roles.map((r) => (
-          <li key={r.username}>
-            {r.username}: {r.role}
-          </li>
-        ))}
-      </ul>
+      <RolesList />
     </div>
   );
 }
